+++ openvino_xai/explainer/explanation.py
@@ -7,9 +7,11 @@ from pathlib import Path
 from typing import Any, Dict, List
 
 import cv2
+import matplotlib.pyplot as plt
 import numpy as np
 
 from openvino_xai.common.parameters import Task
+from openvino_xai.common.utils import logger
 from openvino_xai.explainer.utils import (
     convert_targets_to_numpy,
     explains_all,
@@ -135,22 +137,147 @@ class Explanation:
                 raise ValueError("Provided targer index {targer_index} is not available among saliency maps.")
         return target_indices
 
-    def save(self, dir_path: Path | str, name: str | None = None) -> None:
-        """Dumps saliency map."""
+    def save(
+        self,
+        dir_path: Path | str,
+        prefix: str = "",
+        postfix: str = "",
+        confidence_scores: Dict[int, float] | None = None,
+    ) -> None:
+        """
+        Dumps saliency map images to the specified directory.
+
+        Allows flexibly name the files with the prefix and postfix.
+        {prefix} + target_id + {postfix}.jpg
+
+        Also allows to add confidence scores to the file names.
+        {prefix} + target_id + {postfix} + confidence.jpg
+
+        save(output_dir) -> aeroplane.jpg
+        save(output_dir, prefix="image_name_target_") -> image_name_target_aeroplane.jpg
+        save(output_dir, postfix="_class_map") -> aeroplane_class_map.jpg
+        save(
+            output_dir, prefix="image_name_", postfix="_conf_", confidence_scores=scores
+        ) -> image_name_aeroplane_conf_0.85.jpg
+
+        Parameters:
+            :param dir_path: The directory path where the saliency maps will be saved.
+            :type dir_path: Path | str
+            :param prefix: Optional prefix for the saliency map names. Default is an empty string.
+            :type prefix: str
+            :param postfix: Optional postfix for the saliency map names. Default is an empty string.
+            :type postfix: str
+            :param confidence_scores: Dict with confidence scores for each class index. Default is None.
+            :type confidence_scores: Dict[int, float] | None
+
+        """
+
         os.makedirs(dir_path, exist_ok=True)
-        save_name = name if name else ""
-        for cls_idx, map_to_save in self._saliency_map.items():
+
+        template = f"{prefix}{{target_name}}{postfix}{{conf_score}}.jpg"
+        for target_idx, map_to_save in self._saliency_map.items():
+            conf_score = ""
             map_to_save = cv2.cvtColor(map_to_save, code=cv2.COLOR_RGB2BGR)
-            if isinstance(cls_idx, str):
-                cv2.imwrite(os.path.join(dir_path, f"{save_name}.jpg"), img=map_to_save)
-                return
+            if isinstance(target_idx, str):
+                target_name = "activation_map"
+            elif self.label_names and isinstance(target_idx, np.int64):
+                target_name = self.label_names[target_idx]
             else:
-                if self.label_names:
-                    target_name = self.label_names[cls_idx]
-                else:
-                    target_name = str(cls_idx)
-            image_name = f"{save_name}_target_{target_name}.jpg" if save_name else f"target_{target_name}.jpg"
+                target_name = str(target_idx)
+
+            if confidence_scores and target_idx in confidence_scores:
+                conf_score = f"{confidence_scores[int(target_idx)]:.2f}"
+
+            image_name = template.format(target_name=target_name, conf_score=conf_score)
             cv2.imwrite(os.path.join(dir_path, image_name), img=map_to_save)
+
+    def plot(
+        self,
+        targets: np.ndarray | List[int | str] | None = None,
+        backend: str = "matplotlib",
+        max_num_plots: int = 24,
+        num_columns: int = 4,
+    ) -> None:
+        """
+        Plots saliency maps using the specified backend.
+
+        This function plots available saliency maps using the specified backend. Targets to plot
+        can be specified by passing a list of target class indices or names. If a provided class is
+        not available among the saliency maps, it is omitted.
+
+        Args:
+            targets (np.ndarray | List[int | str] | None): A list or array of target class indices or names to plot.
+                By default, it's None, and all available saliency maps are plotted.
+            backend (str): The plotting backend to use. Can be either 'matplotlib' (recommended for Jupyter)
+                or 'cv' (recommended for Python scripts). Default is 'matplotlib'.
+            max_num_plots (int): Max number of images to plot. Default is 24 to avoid memory issues.
+            num_columns (int): Number of columns in the saliency maps visualization grid for the matplotlib backend.
+        """
+
+        if targets is None or explains_all(targets):
+            checked_targets = self.targets
+        else:
+            target_indices = get_target_indices(targets, self.label_names)
+            checked_targets = []
+            for target_index in target_indices:
+                if target_index in self.saliency_map:
+                    checked_targets.append(target_index)
+                else:
+                    logger.info(f"Provided class index {target_index} is not available among saliency maps.")
+
+        if len(checked_targets) > max_num_plots:
+            logger.warning(
+                f"Decrease the number of plotted saliency maps from {len(checked_targets)} to {max_num_plots}"
+                " to avoid the memory issue. To avoid this, increase the 'max_num_plots' argument."
+            )
+            checked_targets = checked_targets[:max_num_plots]
+
+        if backend == "matplotlib":
+            self._plot_matplotlib(checked_targets, num_columns)
+        elif backend == "cv":
+            self._plot_cv(checked_targets)
+        else:
+            raise ValueError(f"Unknown backend {backend}. Use 'matplotlib' or 'cv'.")
+
+    def _plot_matplotlib(self, checked_targets: list[int | str], num_cols: int) -> None:
+        """Plots saliency maps using matplotlib."""
+        num_rows = int(np.ceil(len(checked_targets) / num_cols))
+        _, axes = plt.subplots(num_rows, num_cols, figsize=(5 * num_cols, 6 * num_rows))
+        axes = axes.flatten()
+
+        for i, target_index in enumerate(checked_targets):
+            if self.label_names and isinstance(target_index, np.int64):
+                label_name = f"{self.label_names[target_index]} ({target_index})"
+            else:
+                label_name = str(target_index)
+
+            map_to_plot = self.saliency_map[target_index]
+
+            axes[i].imshow(map_to_plot)
+            axes[i].axis("off")  # Hide the axis
+            axes[i].set_title(f"Class {label_name}")
+
+        # Hide remaining axes
+        for ax in axes[len(checked_targets) :]:
+            ax.set_visible(False)
+
+        plt.tight_layout()
+        plt.show()
+
+    def _plot_cv(self, checked_targets: list[int | str]) -> None:
+        """Plots saliency maps using OpenCV."""
+        for target_index in checked_targets:
+            if self.label_names and isinstance(target_index, np.int64):
+                label_name = f"{self.label_names[target_index]} ({target_index})"
+            else:
+                label_name = str(target_index)
+
+            map_to_plot = self.saliency_map[target_index]
+            map_to_plot = cv2.cvtColor(map_to_plot, cv2.COLOR_BGR2RGB)
+
+            cv2.imshow(f"Class {label_name}", map_to_plot)
+            cv2.waitKey(0)
+        cv2.destroyAllWindows()
 
 
 class Layout(Enum):