+++ tests/unit/explanation/test_explanation.py
@@ -1,9 +1,11 @@
 # Copyright (C) 2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
+import logging
 import os
 
 import numpy as np
+import pytest
 
 from openvino_xai.explainer.explanation import Explanation
 from tests.unit.explanation.test_explanation_utils import VOC_NAMES
@@ -41,23 +43,37 @@ class TestExplanation:
         save_path = tmp_path / "saliency_maps"
 
         explanation = self._get_explanation()
-        explanation.save(save_path, "test_map")
-        assert os.path.isfile(save_path / "test_map_target_aeroplane.jpg")
-        assert os.path.isfile(save_path / "test_map_target_bird.jpg")
+        explanation.save(save_path, prefix="image_name_")
+        assert os.path.isfile(save_path / "image_name_aeroplane.jpg")
+        assert os.path.isfile(save_path / "image_name_bird.jpg")
 
         explanation = self._get_explanation()
         explanation.save(save_path)
-        assert os.path.isfile(save_path / "target_aeroplane.jpg")
-        assert os.path.isfile(save_path / "target_bird.jpg")
+        assert os.path.isfile(save_path / "aeroplane.jpg")
+        assert os.path.isfile(save_path / "bird.jpg")
 
         explanation = self._get_explanation(label_names=None)
-        explanation.save(save_path, "test_map")
-        assert os.path.isfile(save_path / "test_map_target_0.jpg")
-        assert os.path.isfile(save_path / "test_map_target_2.jpg")
+        explanation.save(save_path, postfix="_class_map")
+        assert os.path.isfile(save_path / "0_class_map.jpg")
+        assert os.path.isfile(save_path / "2_class_map.jpg")
+
+        explanation = self._get_explanation()
+        explanation.save(save_path, prefix="image_name_", postfix="_map")
+        assert os.path.isfile(save_path / "image_name_aeroplane_map.jpg")
+        assert os.path.isfile(save_path / "image_name_bird_map.jpg")
+
+        explanation = self._get_explanation()
+        explanation.save(save_path, postfix="_conf_", confidence_scores={0: 0.92, 2: 0.85})
+        assert os.path.isfile(save_path / "aeroplane_conf_0.92.jpg")
+        assert os.path.isfile(save_path / "bird_conf_0.85.jpg")
 
         explanation = self._get_explanation(saliency_maps=SALIENCY_MAPS_IMAGE, label_names=None)
-        explanation.save(save_path, "test_map")
-        assert os.path.isfile(save_path / "test_map.jpg")
+        explanation.save(save_path, prefix="test_map_")
+        assert os.path.isfile(save_path / "test_map_activation_map.jpg")
+
+        explanation = self._get_explanation(saliency_maps=SALIENCY_MAPS_IMAGE, label_names=None)
+        explanation.save(save_path, prefix="test_map_", postfix="_result")
+        assert os.path.isfile(save_path / "test_map_activation_map_result.jpg")
 
     def _get_explanation(self, saliency_maps=SALIENCY_MAPS, label_names=VOC_NAMES):
         explain_targets = [0, 2]
@@ -67,3 +83,39 @@ class TestExplanation:
             label_names=label_names,
         )
         return explanation
+
+    def test_plot(self, mocker, caplog):
+        explanation = self._get_explanation()
+
+        # Invalid backend
+        with pytest.raises(ValueError):
+            explanation.plot(backend="invalid")
+
+        # Plot all saliency maps
+        explanation.plot()
+        # Matplotloib backend
+        explanation.plot([0, 2], backend="matplotlib")
+        # Targets as label names
+        explanation.plot(["aeroplane", "bird"], backend="matplotlib")
+        # Plot all saliency maps
+        explanation.plot(-1, backend="matplotlib")
+        # Update the num columns for the matplotlib visualization grid
+        explanation.plot(backend="matplotlib", num_columns=1)
+
+        # Class index that is not in saliency maps will be ommitted with message
+        with caplog.at_level(logging.INFO):
+            explanation.plot([0, 3], backend="matplotlib")
+        assert "Provided class index 3 is not available among saliency maps." in caplog.text
+
+        # Check threshold
+        with caplog.at_level(logging.WARNING):
+            explanation.plot([0, 2], backend="matplotlib", max_num_plots=1)
+
+        # CV backend
+        mocker.patch("cv2.imshow")
+        mocker.patch("cv2.waitKey")
+        explanation.plot([0, 2], backend="cv")
+
+        # Plot activation map
+        explanation = self._get_explanation(saliency_maps=SALIENCY_MAPS_IMAGE, label_names=None)
+        explanation.plot()